Renamed GroupDetails.ts to GroupHeadings.ts, for sole remaining class
